Swaps the `app-root-path` package for the `process.env.PWD` variable to get the root directory
This should be more reliably for getting the root directory of the project where the process is being run. This also fixes where devs couldn't use this package if it were install from a local path. For Windows PCs where `process.env.PWD` is not set, it should default to using `process.cwd()` instead.